navbar only shows up when user is logged in, otherwise landing page redirects to landing page when user logs out
