+++ src/client/App.js
@@ -1,34 +1,51 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import { BrowserRouter, Routes, Route, Link } from 'react-router-dom';
-import { Counter } from './Components/misc/Counter/Counter';
-import GetResource from './Components/GetResource/GetResource';
 import Home from './Components/Home/Home';
 import CreateCard from './Components/CreateCard/CreateCard';
 import Navbar from './Components/Navbar/Navbar';
 import FlashCard from './Components/ShowCard/FlashCards';
 import LandingPage from './Components/LandingPage/LandingPage';
-
-// alina testing
-
+import { useSelector, useDispatch } from 'react-redux';
+import { setEmail } from './Redux/slices/userSlice';
+import axios from 'axios';
 const App = () => {
   const leftItems = {
-    home: '/',
+    home: '/library',
   };
+
+  // On first render, get user data
+  const dispatch = useDispatch();
+  useEffect(() => {
+    const response = axios({
+      method: 'get',
+      withCredentials: true,
+      url: 'http://localhost:8080/auth/user',
+    }).then((res) => {
+      if (res.data) {
+        dispatch(setEmail(res.data.email));
+      }
+    });
+  }, []);
+  const email = useSelector((state) => state.user.email);
+
   return (
     <>
-      <BrowserRouter>
-        {/* Component 'Navbar' must be placed within browser router so that navbar links work */}
-        <Navbar leftItems={leftItems} />
-        <div>
-          <Routes>
-            <Route exact path="/" element={<LandingPage />} />
-            <Route exact path="/library" element={<Home />} />
-            <Route exact path="/api/:resource/:id" element={<GetResource />} />
-            <Route exact path="/createCard" element={<CreateCard />} />
-            <Route exact path="/flashcard/:id" element={<FlashCard />} />
-          </Routes>
-        </div>
-      </BrowserRouter>
+      {!email ? (
+        <LandingPage />
+      ) : (
+        <BrowserRouter>
+          {/* Component 'Navbar' must be placed within browser router so that navbar links work */}
+          <Navbar leftItems={leftItems} />
+          <div>
+            <Routes>
+              <Route exact path="/" element={<LandingPage />} />
+              <Route exact path="/library" element={<Home />} />
+              <Route exact path="/createCard" element={<CreateCard />} />
+              <Route exact path="/flashcard/:id" element={<FlashCard />} />
+            </Routes>
+          </div>
+        </BrowserRouter>
+      )}
     </>
   );
 };
